move UpgradeRequiredComponent to Businesses namespace
