+++ app/components/businesses/upgrade_required_component.html.erb
