+++ app/views/cold_messages/new.html.erb
@@ -35,7 +35,7 @@
         </div>
       </div>
     <% else %>
-      <%= render UpgradeRequiredComponent.new(current_user) %>
+      <%= render Businesses::UpgradeRequiredComponent.new(current_user) %>
     <% end %>
   </div>
 </div>
